// frontend/src/lib/api/model/projectResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Roxai API
 * OpenAPI spec version: 1.0.0
 */
import type { ProjectStatus } from './projectStatus';
import type { ProjectResponseThumbnailUrl } from './projectResponseThumbnailUrl';
import type { ProjectResponseFolderId } from './projectResponseFolderId';

export interface ProjectResponse {
  id: string;
  title: string;
  status: ProjectStatus;
  thumbnail_url?: ProjectResponseThumbnailUrl;
  folder_id?: ProjectResponseFolderId;
  created_at: string;
  updated_at: string;
}