// frontend/src/lib/api/model/projectStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Roxai API
 * OpenAPI spec version: 1.0.0
 */

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ProjectStatus = {
  draft: 'draft',
  generating: 'generating',
  done: 'done',
  failed: 'failed',
} as const;